Add adjustMinimumWidths implementation. Rename DataCellWidths to ColumnWidths.
